flow: performance optimization -- avoid O(N^2)

+++ kzen-auto-common/src/commonMain/kotlin/tech/kzen/auto/common/paradigm/flow/model/exec/ActiveVertexModel.kt
@@ -6,7 +6,7 @@ class ActiveVertexModel(
         var message: Any?,
         val remainingBatch: MutableList<Any>,
         var streamHasNext: Boolean,
-        var epoch: Long,
+        var epoch: Int,
 
         // TODO: factor out to be visual-only for performance?
         var error: String?
